+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/groq
 
+## 3.0.1
+
+### Patch Changes
+
+- Updated dependencies [29264a3]
+  - @ai-sdk/provider-utils@4.0.1
+
 ## 3.0.0
 
 ### Major Changes